requirements.txt
streamlit
google-genai
PymuPDF
weasyprint
Markdown-It-Py